+++ src/main/java/com/cloudogu/scm/review/comment/service/BasicComment.java
@@ -29,9 +29,9 @@ import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlRootElement;
 import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
+import java.io.Serializable;
 import java.time.Instant;
 import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
@@ -41,7 +41,7 @@ import static java.util.Collections.unmodifiableSet;
 
 @XmlRootElement(name = "comment")
 @XmlAccessorType(XmlAccessType.FIELD)
-public abstract class BasicComment implements Cloneable {
+public abstract class BasicComment implements Cloneable, Serializable {
 
   private String id;
   private String comment;